# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel v11 auth API's",
        "version": "1.0"
    },
    "paths": {
...
        "/api/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get filtered news articles",
                "description": "Fetch news articles based on optional filters like keyword, date, category, source, and pagination or everything.",
                "operationId": "9bfe744cf19dcb58e9ab9677c29ee6c7",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search keyword for news title or description",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Election results"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Publication date of the news article",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-11-13"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category of the news article",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "technology",
                                "business",
                                "sports",
                                "health",
                                "entertainment"
                            ],
                            "example": "technology"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Source of the news article",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "BBC News"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of filtered news articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Latest Technology Trends"
                                                    },
                                                    "source": {
                                                        "type": "string",
                                                        "example": "BBC News"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "technology"
                                                    },
                                                    "author": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "published_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-13T10:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 20
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 100
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid filter provided"
                                        },
                                        "message": {
                                            "properties": {
                                                "keyword": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The keyword must be a string."
                                                    }
                                                },
                                                "date": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The date is not a valid date."
                                                    }
                                                },
                                                "category": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The selected category is invalid."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "An error occurred while fetching news articles"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/news/{news}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get a specific news article",
                "description": "Fetch a single news article by its news Id.",
                "operationId": "4c09fa11f86d11aa99aefdc9f85c1237",
                "parameters": [
                    {
                        "name": "news",
                        "in": "path",
                        "description": "ID of the news article to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "News article found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Latest Technology Trends"
                                        },
                                        "source": {
                                            "type": "string",
                                            "example": "BBC News"
                                        },
                                        "category": {
                                            "type": "string",
                                            "example": "technology"
                                        },
                                        "author": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "published_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-13T10:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "News article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "News article not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "An error occurred while fetching news"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}